Don't log confusing lines from parcel (#1078)
Stops logging `Server running at http://localhost:1234` which is super
confusing because that port number is only in the docker container. It
also suppresses logging of `✨  Built in` until the nginx server is up
and running because it can give folks the false impression that that
docs are ready when it is really just saying the web resources are
ready.