Prevent jumping when you pick an alternative lang (#1095)
When you pick an alternative language the page has to reflow because the
alternatives have different vertical size. This reflow looks like a big
"jump", especially when you pick the language lower down on the page.
This prevents the "jump" by storing the location of the picker before
the reflow, updating the styles to cause the reflow, grabbing the new
location, and then scrolling the window the difference between the new
and old location. This keeps the picker, which is usually right under
your mouse when you change the language, from jumping. This "feels"
smooth!