Style: Duplicate styles from next.js (#1101)
The header and footer and delivering a bunch of `#guide` specific styles
that we'd prefer to have control over ourselves. We keep up with the
public site for the most part, but we'd like to do it on our schedule.
This duplicates the `#guide` styles from the public site on our side so
it can safely drop them.