Automatically update version drop downs when we add or remove a version (#1062)
I was inspired by #1055 to see if I could make it so that we don't have
to run a build with `--rebuild` when we release. I failed. But I made
progress on one front so I think it is worth merging. This change makes
it so that we rebuild the "version" drop down if we add or remove versions
from a book. This'll probably allow us to skip running `--rebuild` when we
add new branches but it isn't enough to stop us from having to run a
`--rebuild` with every release because:

1. It doesn't rebuild the version drop down if we only switch the current
version to another version.
2. It won't rebuild the headers that tell you that you are reading docs
for an old version or for an unreleased version when we switch the current
version to another version.

But it *is* a start.