Copy images from unordeed and definition lists (#1093)
Entirely reworks how we copy images by adding a new type of extension
to Asciidoctor: delegating converters. Adding a new converter isn't
documented in their user's manual but their code makes it clear that they
expect folks to plug in customized extensions.

These make copying images *much* simpler because it allows us to
intercept the call for inline images correctly *and* it lets us get at
the attribute values.

This mechanism is also useful for customizing how we generate output
and this PR uses it to clean up one particularly gnarly output mechanism
that was also getting in the way of copying images. It is now much more
readable on both the output generation and image copying side.