Move the repos dir config to parameter (#1117)
Removes the `paths.repos` config entry from the conf file because we
rarely want to change it outside of tests. Adds a command line parameter
to change the path (`--reposcache`) and uses it in the tests.

This sort of accidentally fixes a problem where running `--self-test`
will trash the contents of `.repos/target_repo.git`, forcing subsequent
`--all` builds on the same machine to fetch it.