Shrink diffs for sitemap.xml (#1110)
We generate `sitemap.xml` with an entry for every page in every
"current" book. Before this change we'd set the last modified time to
"now" for every single page, every time we rebuild the books. This
changes it so that we only change the last modified time when we modify
the file which is how you are supposed to use `sitemap.xml` anyway.

Beyond that, it sorts the entries in the sitemap before building it so
it doesn't "jump arround" on every commit.

These two combine to significantly shrink the size of small diffs which
should make working with the docs just a little bit faster.

Closes #684