Skip new --sub_dirs unless all sources are present
When we added a new branch to the list of books last week we broke the
docs PR tests for books that were involved in more than one repo. The
build saw `--sub_dir` for *that* repo and no other history so it tried
to build those multi-source books with just the `--sub_dir`ed repo. This
didn't work because we were missing a ton of files.

This detects when `--sub_dir` is used for a new source and skips
building the book *unless* all of the sources are new *and*
`--sub_dir`ed. This means we'll still build, say, the Elasticsearch
reference on Elasticsearch PRs even if the book is new but we won't
build, say, the stack overview. So we lose test coverage here but in
exchange we don't fail builds.

The more paranoid option would be to "fail closed" and log a more
helpful message but fail the build until the "global" docs build has
run. This doesn't seem like the right option because, frankly, we'd
prefer to keep merging a little smoother on release day at the cost of
failing the global docs build a little more. We are used to some trouble
with the global docs build on release day.