Replace perl template with js template (#1155)
This replaces the perl templating code with javascript that does almost
exactly the same thing. The implementations are almost word for word the
same. This is ever so slightly slower because the template has to be
initalized once per book instead of once and because of the fork
overhead of starting a new node process. It has the distinct advantage
of being much simpler to embed in something like the preview service.