Allow selecting headings
Previously our headings were "covered up" by a funny anchor tag that we
used to make it possible to link to portions of the page. This made them
impossible to select. This changes headings so they are *part of* the
anchor tag. Text inside an anchor tag isn't selectable on its own but
you *can* select it by dragging from outside of the tag inwards. This is
"how browsers work" so folks should be used to this. Hopefully making
the text part of the anchor tag will make it much more obvious what is
going on.

Closes #277