Template on the fly in preview (#1191)
This makes the preview service template html pages on the fly.

A few things had to happen to make this go:
1. Start writing everything that we need to render the tempelate on the
fly into built-docs, specifically the template and the language.
2. Only use the template if it is in the built-docs branch we're trying
to render. Branches without the template serve from the "html" directory
like we do in production.
3. Rework error handling for the git operations.
4. Move the git operations into their own file and unit test them.
5. Rework some other error handling after I found out the `.pipe`
doesn't propagate errors in node.

The performance cost of apply the template on the fly is something like
20 milliseconds which doesn't feel to bad because we're running six git
operations instead of two.