Fix toc in preview (#1217)
Fix the table of contents in the preview and `--doc --open` modes.
Without this the table of contents just *doesn't* show up. With this we
continue to perform client side merging of the toc into the page.

Now that we're templating on the fly it is tempting to merge the toc at
template time, but that won't work in production because we're not
templating on the fly in production. Yet.