Air gapped docs always serve master (#1235)
I'd configured the air gapped docs so if you access them as
`localhost/blah` they'd always give you the master branch. But I still
allowed you to grab other branches with `foo.localhost/blah`. The
trouble with that is that folks will tend to want to have `.`s in the
hostname when they deploy it.

This locks the branch to always be `master` regardless of the host
header.