Check warnings from asciidoctor before next steps
Before this commit we bundled executing `asciidoctor`, `xmllint`, and
`xsltproc` together, checking warnings after the whole process
succeedes. If any step in the process fails we'd log the errors for
*that* *step* and exit non-0. The trouble with this behavior is that
we'd throw out the warnings from the previous steps. This is especially
bad because the warnings from `asciidoctor` are *much* easier to debug
that errors from `xmllint`.

This commit causes us to check the warnings after each step. If there
are any warnings we log them and exit non-0. This is will effectively
prefer warnings from asciidoctor if there are any which should make
debugging problems like bad links *much* simpler.