Switch the edit_me link implementation (#1239)
Switches the edit_me link implementation from a TreeProcessor to a
DelegatingConverter. This is nice because, eventually, we'll want to
generate the edit_me links differently if we're converting to html
instead of docbook. Converters are the place where we can make that
call.

Also I can remove some rubocop suppressions which is nice.