Speed up integration tests (#1252)
Before: 10m47.920s
After:   6m1.201s

The integration tests have felt slower to me so I did a little digging.
`top` told me that git was pretty busy during the integration tests and
it dawned on me that we put the fonts in the built-docs repo now and
they are pretty big.

This speeds up the tests by creating a "shell" repository before most
tests that already has the fonts in it and then clonging the "target"
repository from that shell before every test. That way we only have to
pay the cost of `git add`ing the font files one time.