Make sure we don't rebuild at a bad time (#1257)
It is important that changes to the docs repo don't trigger a rebuild
when `--keep_hash` is specified or else every time we change this
file we'll rebuild all books in every PR build. That is a waste of
time and a potential source of spurious errors.

This PR does exactly that: it adds tests that make sure we don't rebuild
and it makes a small change to the initialization fo the docs repo to
make the test pass.