Fancy --sub_dir and --keep_hash
This adds a fancy interaction between `--sub-dir` and `--keep_hash` that
is meant to make sure we build *exactly* the right docs for pull
requests.

When we first deploy pull request builds we targeted them at the tip of
the pull request branch. This caused trouble because the tip of the pull
request branch was often behind the last successful build. Building the
docs against that would often cause spurious failure from out of date
links. Merging `master` into the PR branch usually fixed it but it was
annoying.

We fixed this problem by switching our target to the result of merging
the pull request branch into the target branch. This was *better*
because it had fewer spurious failures, but often caused us to think
that the docs changes were larger than they actually were because the
target branch often had changes in it that hadn't yet been picked up by
the docs build. This was annoying because the `diff`s generated by the
docs builds had changes that weren't part of the PR. Which made it
difficult to track down what actually changed.

This solves the problem by first merging the `--sub_dir`ed branch into
the last successful docs build. Thus we only build *exactly* what is in
the PR. We trigger this behavior by checking for both the `--keep_hash`
and `--sub_dir` flags *and* checking for outstanding changes in the
`--sub_dir`ed branch.

This behavior might end up being slower sometimes because merging can be
expensive, but it is usually less expensive than building a book we
don't have to build. So maybe it'll all come out in the wash.