Add APM app attribute
This PR adds a new attribute for the APM app documentation: `apm-app-ref`. Right now, this reference points to the Kibana docs. Over the next few weeks I'll slowly be swapping our Kibana links to use this attribute. Then, when the APM app documentation moves out of Kibana, I'll simply need to adjust this attribute to point to the right location.