Redirect slightly more
When you make a request to a location outside of the `guide` we should
redirect to `www.elastic.co` *unless*:
1. We have the file (then we should just send it to you).
2. You made the request for air gapped docs (we should 404).

This implements that by adding a `try_files` block to the `/` arm of our
nginx config which forwards requests for missing files right into the
preview application. The preview application checks if you are asking
for the air gapped deploy or not and performs the redirect or not.

Relates to #1300