Fix error reporting when book depends on docs (#1292)
When a book fails to build we try to list the latest commits from each
of the sources of the book. When one of the sources of that book is the
docs repo *and* the docs repo was cloned with `--reference` then we were
failing to list the source and crashing. This fixes that by mounting the
`--reference`d repo.

Co-Authored-By: Josh Mock <josh@joshmock.com>