Simplify open_in_widget
I've been wanting to remove our fairly complex `TreeProcessorScaffold`
for a while now because it mirror's Asciidoctor's `Conveter` except it
doesn't properly handle attributes. This migrates `open_in_widget` off
of the scaffold to a converter. This also has the advantage of giving a
clear path to implement the converter if we target html instead of
docbook, which we'd very much like to do.