Upgrade asciidoctor (#1358)
We need to pick up the fix for
https://github.com/asciidoctor/asciidoctor/issues/3089 so we need to
upgrade. These are the changes required to upgrade:

* Add our own docbook 4.5 converter because Asciidoctor 2 dropped
support for docbook 4.5. We'll drop support for docbook entirely in a
follow up so there is no point in upgrading to docbook 5.
* Rework converter methods to be `convert_{name}` instead of `{name}`.
* Replace `with_format` with `format`.
* `node.attr` no longer delegates missing attributes to the document.
You now be explict if you want an attribute from the document and do
`node.document.attr`.
* Asciidoctor no longer `WARN`s when it sees an invalid reference so we
can't fail on one without xmllint. Again. This is sad because the error
reporting is much worse.