Emulate docbook chunking
We'd very much like to avoid having to use docbook to generate our HTML
because it is slow and difficult to customize. This creates the
beginnings of an Asciidoctor plugin that emulates docbook's chunking.
The plan is to enable it on the perl side with a command line flag and a
flag in conf.yaml in a follow up change.