Clean ignores for "normal" docker image (#1398)
Our "normal" docker image was including context files needed for the
`preview` docker image because I didn't know how to make separate
`.dockerignore` files for each `Dockerfile`. I know now. So this fixes
up the main `.dockerignore` file.

You'd expect this to speed up building the docker image but it doesn't
seem to do so. It is still nice though....