Shrink production images
This shrinks the build, preview, and air gapped images by moving the
test dependencies out of them. This is also nice because changing test
dependencies won't cause us to have to rebuild the production images.
The cost is that it'll be a little slower to run the tests the first
time. This isn't a problem locally, but will slow down CI. We can
mostly mitigate that in a follow up change though.

Before:
```
-rw------- 1 manybubbles manybubbles 1.1G Nov 12 19:06 air_gapped_with_test.tar
```

After:
```
-rw------- 1 manybubbles manybubbles 930M Nov 12 19:01 air_gapped_no_test.tar
```