Progress building chunked docs (#1404)
This hooks the `chunker` extension added in #1390 to the docs build when
you do a multi-page book with `--direct_html`. Important: it crashes!
But I'll fix it with mostly ruby hacking and this is more perl stuff and
it is fairly selef contained. So I figure I should fix it in a follow up
change.