Speed up tests
Builds the testing docker images during the packer cached step of the
process that builds the images that run the docs build. So *usually*
we'll be able to use the cached copy. Which speeds up the tests.

As a bonus this builds the html diff tool during the tests and packer
cache phase. This both makes sure that we have the image if we need it
*and* makes sure that we don't break being about to build the image when
we make changes.