Create a "live" branches config for books
We've wanted for a while to mark old versions of books as "noindex" and,
probably, to remove the `edit_me` links from them. This creates a config
entry for `conf.yaml` to mark branches as "live". If this list is
specified then everything *not* in the list will be marked as `noindex`.
In the future we might remove the edit links from it was well.

This applies that config change to the installation and upgrade guide,
just to apply it to something for testing. I think we'll want to apply
it to most books eventually.