Create docbook workalike plugin (#1420)
Creates a plugin to "work like" Elastic's docbook customizations. The
idea here is to replace docbook with as close to 0 diffs as possible.
But in some cases, that isn't worth it. So this also alters `html_diff`
to start to normalize all of the diffs that we don't care about. Right
now there is only one, but I'm sure we'll have a hand full.

The actual asciidoctor plugin munges the header and title into the
correct form with string manipulation. I don't feel *too* bad about that
because:
1. It is in the grand tradition of bubble gum and duct tape that holds
all of the important bits of the internet together.
2. We'll be able to *remove* it piece by piece once we're done. It'll be
*much* easier to visually verify that removing each piece is safe if we
can do it one by one.
3. Further work on the plugin will *probably* be cleaner because the
other parts of the document are generated much more precisely so we
*should* be able to override more carefully.