Improve direct html generation
This improves our plugin for direct html generation in Asciidoctor to
the point where the diff for the perl client's docs is empty. Now, that
isn't as big a feat as it sounds because:
1. Most of the perl client's docs are on CPAM so the docs we actually
build are pretty small.
2. I pull a few tricks with `html_diff` to ignore a few changes that
don't *look* like they make any visual difference to me.

Even with those two caveates, this is pretty good progress. It switches
the perl client to direct html, because, well, I did all the work to
make sure the diff is empty.