Clean up yaml generation in tests (#1422)
The integration tests used a lot of code to properly generate yaml that
the docs build could parse, mostly because the docs build used an old
yaml parser that had a few bugs. But we recently upgraded the yaml
parser and it looks like it doesn't have the bugs any more! So this
replaces the "by hand" yaml generation with ruby's built in `.to_yaml`
which is much simpler to do.