Improve error messsage when no elastic remote (#1429)
When you build docs with `--doc` and the source files don't live in a
git repo with an Elastic remote we log a line about how we can't find
the Elastic remote. We really only need the remote to build edit urls
so if we don't find it then it is *probably* ok because we just use the
first remote. The edit urls are less realistic but that isn't a big
deal usually.

This modifies the message we log to make it clear that it is just about
the edit links.