Create a substitute for docinfo in direct html (#1438)
We'd very much like to retire docbook, but we use its "docinfo" feature
quite a bit to add extra text to the header of the title page of a few
of our books. Asciidoctor *has* a "docinfo" feature but we can't use it
to emulate docbook's docinfo because the asciidoctor adds the text to
the `<head>` tag while docbook adds it below the `<h1>` tag. This change
causes us to look for files names `-extra-title-page.html` to find text
to add just below the `<h1>`.