Handle single-level TOCs in direct html
This modifies the docbook_compat plugin to add the table of contents to
the same location that docbook adds it and to generate it in the same
style that docbook uses. It only works for a single level, but that
should be fine for now. We'll improve it in further work.