Fix title on direct_html's subpages
`--direct_html` was producing incorrect `<title>`s on subpages. But it
worked in the unit tests! I tracked down the fix to us overwriting the
wrong attribute to force the subpage's title. But again, no luck testing
it in the unit tests.

So this starts testing direct_html in an integration test and adds all
of the right things to catch this.