Add version drop downs to raw (#1457)
Adds the version drop down to the raw docs so they show up in the
preview. It'd be nice if the templating system did this on the fly, but
that is a problem for another day. This is a simple solution that fixes
the preview so we can keep doing other things.