Remove lingering Asciidoc (Python) reference from README (#1479)
The Elastic doc toolchain migrated from the original Python Asciidoc
to Asciidoctor, a re-implementation in Ruby (tracked in #505).

During the migration, the README referenced both Asciidoc and
Asciidoctor. After the migration, #1414 removed all references to the
original Asciidoc and standardized the README on Asciidoctor. However, I
found one lingering reference to the days when both tools were being
used.

Remove a reference to the obsolete `--asciidoctor` command line option
and a note about Asciidoc vs Asciidoctor.