Tolerate direct_html's toc
We use perl code to hack the book's title into the table of contents.
Because we like it there. *Technically* we could use an Asciidoctor
customerization to do it in direct_html but in the spirit of keeping the
changes as small as possible I've just fixed the perl code to tolerate
direct_html output.