Add "other versions" link to versions drop down (#1481)
This adds an "other versions" link to the versions drop downs if we
stripped any versions from the drop down because they aren't "live". We
remove non-"live" versions from the drop down because we don't want
folks accidentally picking up an old version and because the list just
gets too long with all the versions listed. But we still publish all the
old versions in case some unfortunate soul is stuck on that version.
This adds an "other versions" option to the versions drop down so those
folks can find their version.