Update styles for direct_html's screenshots
When we make a screenshot `--direct_html` emits a different class. It is
pretty simple to update the css to pick that class up. Much simpler than
making `--direct_html` emit the docbook classes. So this updates the
styles to render the new class and the html diff tool to ignore the
difference in classes.