Add basic table implementation to direct_html (#1486)
Asciidoctor's default able implementation doesn't line up with
docbook's. So we're going to have to customize it! This adds a very
basic table implemenation that works for the getting started books. And
probably quite a few others. But it is only a start! Tables can get
pretty intricate. We'll get there.