Properly chunk footnotes in direct_html
Asciidoctor allows you to make footnotes and we use them. This makes
sure that they end up on the same page as they are declared. Without it
the footnotes end up on the "index" page which is super confusing.

NOTE: I'm note really making an effort to make this line up perfectly
with docbook's output. These *should* be rare enough that I can verify
them by hand. We really don't use *that* many footnotes.