Add support for abbreviated titles in direct_html
In docbook we can abbreviate the title of a section in the table of
contents by doing something like this:

```
== Intro to XYZ
++++
<titleabbrev>Intro</titleabbrev>
++++

Words.
```

This is *very* docbook. `++++` mean "pass this output directly to the
generated doc" and `<titleabbrev>` is a docbook element.

This change adds support for abbreviating a title of a section in the
table of contents when using `--direct_html` using the following syntax:

```
[id=intro_to_xyz,titleabbrev=Intro]
== Intro to XYZ

Words.
```

It is both shorter and more AsciiDoc friendly. It isn't compatible with
docbook though.

This change *also* adds support for identifying the docbook-style title
abbreviate commands when running with `--direct_html` and having those
set the title abbreviation. In other words: you don't have to change the
old books, but new books that use `--direct_html` should probably use
the new syntax.